+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.model.repository.CartRepository
-import woowacourse.shopping.model.repository.ProductRepository
+import woowacourse.shopping.repository.CartRepository
+import woowacourse.shopping.repository.ProductRepository
 
 class MainViewModel(
     private val productRepository: ProductRepository,
