+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.model.repository.CartRepository
+import woowacourse.shopping.repository.CartRepository
 
 class CartViewModel(
     private val cartRepository: CartRepository,
